+++ structure.go
@@ -1,0 +1,36 @@
+package main
+
+// Task описывает структуру задачи
+// эмулирует работу с полями в базе данных
+type Task struct {
+	ID           string   `json:"id"`
+	Description  string   `json:"description"`
+	Note         string   `json:"note"`
+	Applications []string `json:"applications"`
+}
+
+// tasks хранит список задач
+// эмулиерует работу с базой данных
+var tasks = map[string]Task{
+	"1": {
+		ID:          "1",
+		Description: "Сделать финальное задание темы REST API",
+		Note:        "Если сегодня сделаю, то завтра будет свободный день. Ура!",
+		Applications: []string{
+			"VS Code",
+			"Terminal",
+			"git",
+		},
+	},
+	"2": {
+		ID:          "2",
+		Description: "Протестировать финальное задание с помощью Postmen",
+		Note:        "Лучше это делать в процессе разработки, каждый раз, когда запускаешь сервер и проверяешь хендлер",
+		Applications: []string{
+			"VS Code",
+			"Terminal",
+			"git",
+			"Postman",
+		},
+	},
+}
